LayoutTests/fast/mediastream/RTCPeerConnection-addIceCandidate-expected.txt

Test behavior of RTCPeerConnection.addIceCandidate

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



*** A remote description is needed before a candidate can be added
PASS pc.remoteDescription is null
PASS promise pc.addIceCandidate(new RTCIceCandidate({candidate: 'foo', sdpMid: 0})) rejected with OperationError (DOM Exception 34): Expect line: candidate:<candidate-str>
PASS Remote description set

*** Define sdpMid, badSdpMid, sdpMLineIndex and badSdpMLineIndex for testing
PASS sdpMLineIndex is not badSdpMLineIndex
FAIL Error in promise chain: TypeError: undefined is not an object (evaluating 'pc.getTransceivers()[0].mid')
PASS successfullyParsed is true

TEST COMPLETE
